refactor: change data DBAuthentication implementation name
